MS2 updates
