+++ README.md
@@ -1,1 +1,3 @@
-# ac297r
+# Harvard AC297R Capstone Project
+
+Team Quantum Black working on Drought Prediction
